Agreement owner -> maker/taker
While owner is appropriate terminology for an order given there is sole ownership, an agreement will always have two "owners" which are combinations of maker/taker and floating/fixed.

We should consider how we properly represent this combination on-chain.

Currently "side" refers to whether the maker is fixed or floating side, while maker address and taker address are stored explicitly